Started major rewrite of parsing logic.
Used to be reading a line at a time, which made it awkward to extract specific data
Now reads a token at a time in a stream, so can easily pull out any pieces of data I want (apart from Normals for now)

Also fixed a bug which I didn't know was there, where it didn't add the first triangle to the body.

Only outputting one body for testing for now.